change name userGroup

+++ QuizPractice/Server/smallQuiz/src/main/java/swd392/project/smallquiz/model/entiity/Role.java
@@ -12,7 +12,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    public int roleId;
+    public Long roleId;
     @Column(name = "role_name")
     public String roleName;
 }
